LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-images-to-the-canvas/2d.drawImage.nonfinite.worker-expected.txt


PASS drawImage() with Infinity/NaN is ignored 
